переделал под новую генерецию кода вместо ws/layout -> impl

+++ bin/generate/generateTemplate.js
@@ -3,15 +3,16 @@ Object.defineProperty(exports, "__esModule", { value: true });
 var fs = require('fs');
 exports.generateTemplate = function (dest, template, variables) {
     console.log('generate template to:', dest);
-    var templ = template;
-    templ = templ.replace(/{{ComponentName}}/g, variables.ComponentName);
-    variables.Ws && (templ = templ.replace(/{{Ws}}/g, variables.Ws));
-    variables.Layout && (templ = templ.replace(/{{Layout}}/g, variables.Layout));
-    console.log(templ);
-    try {
-        fs.writeFileSync(dest, templ);
-    }
-    catch (e) {
-        console.error('Cannot write template to folder', dest, template, variables, e);
+    if (!fs.existsSync(dest)) {
+        var templ = template;
+        templ = templ.replace(/{{ComponentName}}/g, variables.ComponentName);
+        variables.Impl && (templ = templ.replace(/{{Impl}}/g, variables.Impl));
+        console.log(templ);
+        try {
+            fs.writeFileSync(dest, templ);
+        }
+        catch (e) {
+            console.error('Cannot write template to folder', dest, template, variables, e);
+        }
     }
 };
